+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 
 ### Fixes
 
+- [#9591](https://github.com/blockscout/blockscout/pull/9591) - Fix duplicated results in `methods-read` endpoint
 - [#9502](https://github.com/blockscout/blockscout/pull/9502) - Add batch_size and concurrency envs for tt token type migration
 - [#9493](https://github.com/blockscout/blockscout/pull/9493) - Fix API response for unknown blob hashes
 - [#9484](https://github.com/blockscout/blockscout/pull/9484) - Fix read contract error